Create README.md
The main objective of the project is to create an online book
store that allows users to search and purchase a book based
on title, author and subject. The selected books are displayed
in a tabular format and the user can order their books online
through credit card payment. The Administrator will have
additional functionalities when compared to the common user.